+++ scripts/module.mjs
@@ -1,9 +1,13 @@
 import { initSettings } from "./settings.mjs"
 import { initQueries } from "./queries.mjs";
-import { initSaveTray } from "./message-handler.mjs"
+import { initSaveTray, readySaveTray } from "./message-handler.mjs"
 
 Hooks.once("init", () => {
   initSettings();
   initQueries();
   initSaveTray();
+});
+
+Hooks.once("ready", () => {
+  readySaveTray();
 });